Refactor InMemoryDatabase class and add account functionality
This commit refactors the InMemoryDatabase class by adding a new method, addAccount, which allows for the addition of new accounts with passwords. The existing setPassword method has been modified to set the initial password for the "Scotty" account. This change improves the functionality of the InMemoryDatabase class and enhances the overall usability of the system.